+++ src/node/desktop/package.json
@@ -3,7 +3,7 @@
   "productName": "RStudio",
   "version": "0.0.1",
   "description": "RStudio",
-  "main": "./dist/main/app.js",
+  "main": "./dist/main/main.js",
   "repository": "https://github.com/rstudio/rstudio",
   "author": "RStudio PBC",
   "license": "AGPL-3.0-only",
@@ -34,8 +34,8 @@
     "watch": "tsc -w",
     "package": "yarn run build && electron-packager ./ --prune --package-manager=yarn --out=./package --overwrite --ignore tsconfig.tsbuildinfo",
     "lint": "eslint ./src ./test",
-    "start": "yarn run build && electron ./dist/main/app.js",
-    "show-version": "yarn run build && electron ./dist/main/app.js --version-json",
+    "start": "yarn run build && electron ./dist/main/main.js",
+    "show-version": "yarn run build && electron ./dist/main/main.js --version-json",
     "test": "electron-mocha -c --config ./test/unit/mocharc.json",
     "testcover": "nyc electron-mocha -c --config ./test/unit/mocharc.json",
     "testwip": "electron-mocha -c -g WIP --config ./test/unit/mocharc.json"
